Update index.html
removed previous change as it did not link up to folder "art home" when changed to "art__home". No current way to change the folders name, might try by deleting it, changing the name on vs and import it again. 

This was a suggestion originally made by html validator but it messed up the images so i am not changingf it again unless absolutly nessecary.